refactor: Moved components into specfic folders
